Fixing failed linting
